+++ filters/utils.go
@@ -1,12 +1,14 @@
 package filters
 
-import "reflect"
+import (
+	"reflect"
+)
 
 // Returns true if input is nil
 func isNil(input interface{}) bool {
 	inputValue := reflect.ValueOf(input)
 	switch inputValue.Kind() {
-	case reflect.Array, reflect.Struct, reflect.Map:
+	case reflect.Array, reflect.Slice, reflect.Struct, reflect.Map:
 		return inputValue.IsNil()
 	default:
 		return false